feat(frontend): [SCRUM-65] Refactor page content into components PageOne and PageTwo, rename hook file to api
